Refactorization and clean-up
