+++ src/main/java/pl/nogacz/chess/application/SaveGame.java
@@ -1,106 +1,117 @@
 package pl.nogacz.chess.application;
 
+import javafx.stage.DirectoryChooser;
+import javafx.stage.Stage;
+import pl.nogacz.chess.application.menu.Statistics;
 import pl.nogacz.chess.board.Board;
 import pl.nogacz.chess.board.Coordinates;
 import pl.nogacz.chess.pawns.PawnClass;
 
 import java.io.*;
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
-import java.util.zip.GZIPInputStream;
-import java.util.zip.GZIPOutputStream;
+import java.util.Map;
+import java.util.zip.CRC32;
 
-/**
- * @author Dawid Nogacz on 10.05.2019
- */
+
 public class SaveGame {
 
-    public boolean isSave() {
-        File tempFile = new File("gameCache/board.dat");
-        return tempFile.exists();
-    }
+    public void save(){
+        DirectoryChooser dirChooser = new DirectoryChooser();
 
-    public void save() {
-        saveBoard();
-        saveChessNotation();
-    }
+        dirChooser.setTitle("Select location");
 
-    private void saveBoard() {
+        File selectedDir = dirChooser.showDialog(new Stage());
+        if(selectedDir != null) {
+            saveBoard(selectedDir);
+            saveChessNotation(selectedDir);
+            saveComputer(selectedDir);
+            saveStatistics(selectedDir);
+        }
+    }
+    private void saveBoard(File dir){
+        String filePath = dir.getAbsolutePath();
+        filePath = filePath + "/board.txt";
+        String buffer = "";
         try {
-            File file = new File("gameCache/board.dat");
-            ObjectOutputStream output = new ObjectOutputStream(new GZIPOutputStream(new FileOutputStream(file)));
-            output.writeObject(Board.getBoard());
-            output.flush();
-            output.close();
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
+            PrintWriter out = new PrintWriter(filePath);
+            HashMap<Coordinates, PawnClass> board = Board.getBoard();
+            for(Map.Entry<Coordinates, PawnClass> entry : board.entrySet()) {
+                String coordinates = entry.getKey().getX() + "," +  entry.getKey().getY() + ",";
+                String man = board.get(entry.getKey()).getColor() + "," + board.get(entry.getKey()).getPawn();
+                String line = coordinates + man;
+                buffer += line;
+                out.println(line);
+            }
+            long crc = computeCRC(buffer.getBytes());
+            out.println(crc);
+            out.close();
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        }
+
+    }
+
+    private void saveChessNotation(File dir) {
+        String filePath = dir.getAbsolutePath();
+        filePath = filePath + "/chessNotation.txt";
+        String buffer = "";
+        try {
+            PrintWriter out = new PrintWriter(filePath);
+            ArrayList<String> list = (ArrayList<String>) ChessNotation.getMovesList();
+           for(int i =0; i<list.size();i++){
+               buffer += list.get(i);
+               out.println(list.get(i));
+           }
+
+           long crc = computeCRC(buffer.getBytes());
+           out.println(crc);
+           out.close();
+        } catch (IOException e) {
+            e.printStackTrace();
         }
     }
 
-    private void saveChessNotation() {
+    private void saveComputer(File dir) {
+        String buffer = "";
         try {
-            File file = new File("gameCache/chessNotation.dat");
-            ObjectOutputStream output = new ObjectOutputStream(new GZIPOutputStream(new FileOutputStream(file)));
-            output.writeObject(ChessNotation.getMovesList());
-            output.flush();
-            output.close();
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
+            String filePath = dir.getAbsolutePath();
+            filePath = filePath + "/computer.txt";
+            PrintWriter writer = new PrintWriter(filePath);
+            writer.println(String.valueOf(Computer.getSkill()));
+            buffer = String.valueOf(Computer.getSkill());
+            long crc = computeCRC(buffer.getBytes());
+            writer.println(crc);
+            writer.flush();
+            writer.close();
+        } catch (IOException e) {
+            e.printStackTrace();
         }
     }
 
-    public void load() {
-        loadBoard();
-        loadChessNotation();
-    }
-
-    private void loadChessNotation() {
+    private void saveStatistics(File dir) {
+        String stats = "Game win:" + Statistics.getGameWin() + "\n" +
+                        "Game loss:" + Statistics.getGameLoss() + "\n"+
+                        "Game draw:" + Statistics.getGameDraw();
         try {
-            Object readObject = readObject(new File("gameCache/chessNotation.dat"));
-
-            if(!(readObject instanceof List)) throw new Exception("Data is not a List");
-
-            List<String> cacheNotation = (List<String>) readObject;
-
-            ChessNotation.setMovesList(cacheNotation);
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
+            String filePath = dir.getAbsolutePath();
+            filePath = filePath + "/statistics.txt";
+            PrintWriter writer = new PrintWriter(filePath);
+            writer.println(stats);
+            long crc = computeCRC(stats.getBytes());
+            writer.println(crc);
+            writer.flush();
+            writer.close();
+        } catch (IOException e) {
+            e.printStackTrace();
         }
     }
 
-    private void loadBoard() {
-        try {
-            Object readObject = readObject(new File("gameCache/board.dat"));
-
-            if(!(readObject instanceof HashMap)) throw new Exception("Data is not a HashMap");
-
-            HashMap<Coordinates, PawnClass> cacheMap = (HashMap<Coordinates, PawnClass>) readObject;
-
-            Board.setBoard(cacheMap);
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
+    private long computeCRC(byte[] byteArr){
+        CRC32 crcmaker = new CRC32();
+        crcmaker.update(byteArr, 0, byteArr.length);
+        return  crcmaker.getValue();
     }
 
-    private Object readObject(File file) {
-        try {
-            ObjectInputStream input = new ObjectInputStream(new GZIPInputStream(new FileInputStream(file)));
-
-            Object readObject = input.readObject();
-            input.close();
-
-            return readObject;
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
-
-        return null;
-    }
-
-    public void remove() {
-        File tempFile = new File("gameCache/board.dat");
-        tempFile.delete();
-        tempFile = new File("gameCache/chessNotation.dat");
-        tempFile.delete();
-    }
 }
